Implementation of topdown checkpoints
