+++ src/lib.rs
@@ -3,9 +3,9 @@
 #![feature(try_blocks)]
 pub mod cli;
 pub mod config;
+pub mod constants;
 pub mod jsonrpc;
 pub mod lotus;
 pub mod manager;
 mod serialization;
 pub mod server;
-pub mod constants;
