+++ src/manager/topdown.rs
@@ -6,9 +6,8 @@ use std::collections::HashSet;
 use std::ops::Deref;
 use std::str::FromStr;
 use std::sync::Arc;
-use std::time::Duration;
 
-use anyhow::Result;
+use anyhow::{Context, Result};
 use cid::Cid;
 use fil_actors_runtime::cbor;
 use fvm_shared::address::Address;
@@ -27,14 +26,14 @@ use crate::jsonrpc::JsonRpcClient;
 use crate::lotus::client::LotusJsonRPCClient;
 use crate::lotus::message::mpool::MpoolPushMessage;
 use crate::lotus::LotusClient;
+use crate::manager::checkpoint::CHAIN_HEAD_REQUEST_PERIOD;
 
-
-/// The frequency at which to check a new chain head.
-const CHAIN_HEAD_REQUEST_PERIOD: Duration = Duration::from_secs(10);
-
-async fn manage_topdown_checkpoints((child, parent): (Subnet, Subnet), stop_notify: Arc<Notify>) -> Result<()> {
+pub async fn manage_topdown_checkpoints(
+    (child, parent): (Subnet, Subnet),
+    stop_notify: Arc<Notify>,
+) -> Result<()> {
     log::info!(
-        "Starting topdown checkpoint manager for (child, parent) subnet pair ({}, {})",
+        "Starting top-down checkpoint manager for (child, parent) subnet pair ({}, {})",
         child.id,
         parent.id
     );
@@ -42,69 +41,73 @@ async fn manage_topdown_checkpoints((child, parent): (Subnet, Subnet), stop_noti
     let child_client = LotusJsonRPCClient::from_subnet(&child);
     let parent_client = LotusJsonRPCClient::from_subnet(&parent);
 
-    // Read the parent's chain head and obtain the tip set CID.
-    log::debug!("Getting parent tipset");
-    let parent_head = parent_client.chain_head().await?;
-    let cid_map = parent_head.cids.first().unwrap().clone();
-    let parent_tip_set = Cid::try_from(cid_map)?;
-
-    // Read the parent's chain head and obtain the topdown checkpoint period.
-    let state = parent_client.ipc_read_gateway_state(parent_tip_set).await?;
-    let period = state.topdown_check_period;
-
-    loop {
+    let result: Result<()> = try {
+        // Read the parent's chain head and obtain the tip set CID.
+        log::debug!("Getting parent tipset");
         let parent_head = parent_client.chain_head().await?;
-
-        let curr_epoch: ChainEpoch = ChainEpoch::try_from(parent_head.height)?;
-        // next epoch to submit.
-        let epoch = checkpoint_epoch(curr_epoch, period);
-
-        // First, we check which accounts are in the validator set. This is done by reading
-        // the parent's chain head and requesting the state at that tip set.
-        let parent_head = parent_client.chain_head().await?;
-        // A key assumption we make now is that each block has exactly one tip set. We panic
-        // if this is not the case as it violates our assumption.
-        // TODO: update this logic once the assumption changes (i.e., mainnet)
-        assert_eq!(parent_head.cids.len(), 1);
         let cid_map = parent_head.cids.first().unwrap().clone();
         let parent_tip_set = Cid::try_from(cid_map)?;
 
-        let subnet_actor_state = parent_client
-            .ipc_read_subnet_actor_state(&child.id, parent_tip_set)
-            .await?;
+        // Read the parent's chain head and obtain the topdown checkpoint period.
+        let state = parent_client.ipc_read_gateway_state(parent_tip_set).await?;
+        let period = state.topdown_check_period;
 
-        let mut validator_set: HashSet<Address, RandomState> = HashSet::new();
-        match subnet_actor_state.validator_set.validators {
-            None => {}
-            Some(validators) => {
-                for v in validators {
-                    validator_set.insert(Address::from_str(v.addr.deref())?);
+        loop {
+            let parent_head = parent_client.chain_head().await?;
+            let curr_epoch: ChainEpoch = ChainEpoch::try_from(parent_head.height)?;
+            // next epoch to submit.
+            let epoch = checkpoint_epoch(curr_epoch, period);
+
+            // First, we check which accounts are in the validator set. This is done by reading
+            // the parent's chain head and requesting the state at that tip set.
+            let parent_head = parent_client.chain_head().await?;
+            assert_eq!(parent_head.cids.len(), 1);
+            let cid_map = parent_head.cids.first().unwrap().clone();
+            let parent_tip_set = Cid::try_from(cid_map)?;
+
+            let subnet_actor_state = parent_client
+                .ipc_read_subnet_actor_state(&child.id, parent_tip_set)
+                .await?;
+
+            let mut validator_set: HashSet<Address, RandomState> = HashSet::new();
+            match subnet_actor_state.validator_set.validators {
+                None => {}
+                Some(validators) => {
+                    for v in validators {
+                        validator_set.insert(Address::from_str(v.addr.deref())?);
+                    }
+                }
+            };
+
+            // For each account that we manage that is in the validator set, we submit a topdown
+            // checkpoint.
+            for account in child.accounts.iter() {
+                if validator_set.contains(account) {
+                    log::debug!("Submitting topdown checkpoint for account {}", account);
+                    submit_topdown_checkpoint(
+                        epoch,
+                        account,
+                        child.id.clone(),
+                        &child_client,
+                        &parent_client,
+                    )
+                    .await?;
                 }
             }
-        };
 
-        for account in child.accounts.iter() {
-            if validator_set.contains(account) {
-                log::debug!("Submitting topdown checkpoint for account {}", account);
-                submit_topdown_checkpoint(
-                    epoch,
-                    account,
-                    child.id.clone(),
-                    &child_client,
-                    &parent_client,
-                )
-                .await?;
+            // Sleep for an appropriate amount of time before checking the chain head again or return
+            // if a stop notification is received.
+            select! {
+                _ = sleep(CHAIN_HEAD_REQUEST_PERIOD) => { continue }
+                _ = stop_notify.notified() => { break }
             }
         }
+    };
 
-        // Sleep for an appropriate amount of time before checking the chain head again or return
-        // if a stop notification is received.
-        if !wait_next_iteration(&stop_notify).await? {
-            return Ok(());
-        }
-    }
-
-    Ok(())
+    result.context(format!(
+        "error in manage_topdown_checkpoints() for subnet pair ({}, {})",
+        parent.id, child.id
+    ))
 }
 
 // Prototype function for submitting topdown messages. This function is supposed to be called each
@@ -128,12 +131,13 @@ async fn submit_topdown_checkpoint<T: JsonRpcClient + Send + Sync>(
 
     // Then, we read from the parent subnet the topdown messages with nonce greater than or equal
     // to the nonce we just obtained.
+    let gateway_addr = Address::from_str(GATEWAY_ACTOR_ADDRESS)?;
     let top_down_msgs = parent_client
-        .ipc_get_topdown_msgs(child_subnet, nonce)
+        .ipc_get_topdown_msgs(&child_subnet, gateway_addr, nonce)
         .await?;
 
     // Finally, we submit the topdown messages to the child subnet.
-    let to = Address::from_str(GATEWAY_ACTOR_ADDRESS)?;
+    let to = gateway_addr;
     let from = *account;
     let topdown_checkpoint = TopDownCheckpoint {
         epoch: parent_epoch,
@@ -148,13 +152,4 @@ async fn submit_topdown_checkpoint<T: JsonRpcClient + Send + Sync>(
     parent_client.mpool_push_message(message).await?;
 
     Ok(())
-}
-
-/// Sleeps for some time if stop_notify is not fired. It returns true to flag that we should move to the
-/// next iteration of the loop, while false informs that the loop should return.
-async fn wait_next_iteration(stop_notify: &Arc<Notify>) -> Result<bool> {
-    select! {
-        _ = sleep(CHAIN_HEAD_REQUEST_PERIOD) => {Ok(true)}
-        _ = stop_notify.notified() => {Ok(false)}
-    }
 }
